Amend initail configuration
