+++ backend/.babelrc
@@ -1,0 +1,9 @@
+{
+    "presets": ["es2015", "stage-2"],
+    "plugins": [
+      ["transform-runtime", {
+        "polyfill": false,
+        "regenerator": true
+      }]
+    ]
+  }
