+++ backend/package.json
@@ -3,36 +3,36 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "babel-node": "babel-node --presets=latest",
-    "build": "babel src -d dist",
-    "start": "node ./bin/www",
-    "dev": "nodemon ./bin/www",
-    "lint": "./node_modules/.bin/eslint **/*.js",
-    "fix": "find ./node_modules/* -mtime +10950 -exec touch {} \\;"
+    "test": "export NODE_PATH=./ --compilers js:babel-register --recursive",
+    "start": "nodemon --exec babel-node -- ./bin/www",
+    "lint": "eslint"
   },
   "dependencies": {
-    "babel-preset-es2015": "^6.24.1",
     "body-parser": "~1.18.2",
+    "compression": "^1.7.2",
     "debug": "~2.6.9",
     "express": "~4.15.5",
     "knex": "^0.14.4",
+    "method-override": "^2.3.10",
     "morgan": "^1.9.0",
     "objection": "^1.0.0",
-    "pg": "^7.4.1"
+    "pg": "^7.4.1",
+    "sqlite3": "^3.1.13"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-latest": "^6.24.1",
-    "eslint": "^4.18.0",
+    "babel-eslint": "^8.2.2",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "eslint": "^4.18.1",
     "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.8.0",
-    "nodemon": "^1.15.1"
-  },
-  "babel": {
-    "presets": [
-      "latest"
-    ]
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-react": "^7.7.0",
+    "mocha": "^5.0.1",
+    "nodemon": "^1.15.1",
+    "rimraf": "^2.6.2"
   }
 }
