--- backend/public/stylesheets/style.css
@@ -1,1 +1,0 @@
-
