+++ backend/src/models/Organisation.js
@@ -1,4 +1,4 @@
-const { Model } = require('objection');
+import { Model } from 'objection';
 
 export default class Organisation extends Model {
   static get tableName() {
