+++ backend/src/routes/api.js
@@ -1,0 +1,12 @@
+import express from 'express';
+import connectDB from '../../config/Db';
+
+const router = express.Router();
+
+connectDB();
+
+router.get('/', (req, res) => {
+  res.json('hello');
+});
+
+module.exports = router;
